moved runworkflowsummary + header tests into workflowsummary folder
