+++ src/app/static/src/tests/components/runWorkflow/workflowSummary/runWorkflowSummary.test.ts
@@ -1,8 +1,8 @@
 import {shallowMount} from "@vue/test-utils";
-import runWorkflowSummary from "../../../js/components/runWorkflow/workflowSummary/runWorkflowSummary.vue"
-import runWorkflowSummaryHeader from "../../../js/components/runWorkflow/workflowSummary/runWorkflowSummaryHeader.vue"
-import {RunWorkflowMetadata} from "../../../js/utils/types";
-import {mockAxios} from "../../mockAxios";
+import runWorkflowSummary from "../../../../js/components/runWorkflow/workflowSummary/runWorkflowSummary.vue"
+import runWorkflowSummaryHeader from "../../../../js/components/runWorkflow/workflowSummary/runWorkflowSummaryHeader.vue"
+import {RunWorkflowMetadata} from "../../../../js/utils/types";
+import {mockAxios} from "../../../mockAxios";
 
 describe(`runWorkflowSummary`, () => {
